Partially implement GET /<key> in client/server
Move some shared functionality to the library like parsing Authorization
headers and HTTP bodies into Boards.

For the server implement the "ever-changing board" for the test key and
receiving the client GET /<key> request. No actual storage/retrieval
exists yet.

For the client implement the GET /<key> for a single key and printing to
the command line.