Add board storage and timestamp parsing
The library now actually implements timestamp parsing is replaced with an
html walker to match the last-modified meta tag and extract the HTTP
formatted time.

The server is updated to check this time and store put boards to disk.
Some checks remain to be implemented such as the difficulty check.

Error handling is still partial.