client: fix edge case in single key get
Getting a single key, that was on on the profile's follower list would
result in fetching the board but not rendering it. This is due to the
"get" function clobbering the follower list with the single key, but the
"render" function using the original follower list to walk the boards.