Refactor disk operations out to store library
This pulls all the "board" storage for the client and the server out to
an encapsulated store library. This increases the modularity of the
codebase and cleans up previously scattered store operations.

Adds a high level of test coverage to the store to increase confidence
in it's accuracy.

Allows future/alternative implementations of board storage such as
sqlite for more indepth metadata/querying.